replace two methods with 3: manhattan A* and euclidean A* and old Dijkstra + add checking diagonal neighbours + now only one function for finding path that takes different heuristic functions (therefore can work as Dijsktra, A* euclidean and A* manhatta)

+++ ClassLibrary/MazeVertex.cs
@@ -3,7 +3,7 @@
     public class MazeVertex
     {
         public List<MazeVertex> neighbours = new List<MazeVertex>();
-        public int cost;
+        public double cost;
         public bool isVisited;
         public int x;
         public int y;
